Separate ArenaWorld entity access
